Rejects composition when the options for a path have different runtime types
During composition validation, a path of the supergraph can have
multiple valid options in subgraphs due to shared fields. When
that happens, this patch validates that the possible runtime types
of the various options are the same. The rational being that
shared fields should resolve the same way in all subgraphs, so
allowing 2 differen solution of shared field to have different set
of possible runtime types is a dangerous proposition: to be correct,
users would have make sure they only ever return value for that field
that are in the intersection of those possible runtime types.
Instead of asking users to reason about this, forcing the sets of
possible runtime types to always be the same (for a given shared field)
ensures any returned value is always in the intersection (since the
intersection of a set with itself is itself).

Fixes #1375